+++ CHANGELOG.md
@@ -25,6 +25,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ### Fixed
+- First-child and last-child selectors are broken
+  ([#293](https://github.com/jjriv/emogrifier/pull/293))
 - Second !important rule needs to overwrite the first one
   ([#292](https://github.com/jjriv/emogrifier/pull/292))