+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 1.1.0 (unreleased)
 
 ### Added
+- Support the attr*=value selector
+  ([#330](https://github.com/jjriv/emogrifier/pull/330))
 - Support the attr^=value selector
   ([#324](https://github.com/jjriv/emogrifier/pull/324))
 - Support the attr~=value selector