+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 1.1.0 (unreleased)
 
 ### Added
+- Support of @import
 - Support the attr*=value selector
   ([#330](https://github.com/jjriv/emogrifier/pull/330))
 - Support the attr$=value selector